tools-s3-sync/requirements.txt
boto3==1.28.23
botocore==1.31.23
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
moto==4.1.14
packaging==23.1
pluggy==1.2.0
pycparser==2.21
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
responses==0.23.3
s3transfer==0.6.1
shellingham==1.5.0.post1
six==1.16.0
typer==0.9.0
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
xmltodict==0.13.0
